+++ server/components/user/index.js
@@ -1,0 +1,5 @@
+/**
+ * Component for users routes
+ * @module user
+ */
+module.exports = require('./user')