+++ server/components/users/index.js
@@ -1,5 +1,0 @@
-/**
- * Component for users routes
- * @module users
- */
-module.exports = require('./users')